task-179. UI changes
