+++ src/Pet.Jira.Application/Articles/Commands/CreateArticle/CreateArticleCommand.cs
@@ -1,0 +1,13 @@
+﻿using MediatR;
+using System;
+
+namespace Pet.Jira.Application.Articles.Commands.CreateArticle
+{
+    public class CreateArticleCommand : IRequest<Guid>
+    {
+        public string Title { get; set; }
+        public string Content { get; set; }
+        public string ImageUrl { get; set; }
+        public string Link { get; set; }
+    }
+}
