+++ src/Pet.Jira.Application/Articles/IArticleRepository.cs
@@ -1,9 +1,9 @@
-﻿using Pet.Jira.Application.Blog.Commands.CreateArticle;
+﻿using Pet.Jira.Application.Articles.Commands.CreateArticle;
 using Pet.Jira.Domain.Entities.Blog;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Pet.Jira.Application.Blog
+namespace Pet.Jira.Application.Articles
 {
     public interface IArticleRepository
     {
