+++ src/Pet.Jira.Application/Articles/Queries/GetArticles/GetArticlesQuery.cs
@@ -1,8 +1,8 @@
 ﻿using MediatR;
-using Pet.Jira.Application.Blog.Dto;
+using Pet.Jira.Application.Articles.Dto;
 using System.Collections.Generic;
 
-namespace Pet.Jira.Application.Blog.Queries.GetArticles
+namespace Pet.Jira.Application.Articles.Queries.GetArticles
 {
     public class GetArticlesQuery : IRequest<IEnumerable<ArticleDto>>
     {
