--- src/Pet.Jira.Application/Blog/Commands/CreateArticle/CreateArticleCommandHandler.cs
@@ -1,24 +1,0 @@
-﻿using MediatR;
-using System;
-using System.Threading;
-using System.Threading.Tasks;
-
-namespace Pet.Jira.Application.Blog.Commands.CreateArticle
-{
-    public class CreateArticleCommandHandler : IRequestHandler<CreateArticleCommand, Guid>
-    {
-        private readonly IArticleRepository _repository;
-
-        public CreateArticleCommandHandler(
-            IArticleRepository repository)
-        {
-            _repository = repository;
-        }
-
-        public async Task<Guid> Handle(CreateArticleCommand request, CancellationToken cancellationToken)
-        {
-            var entity = await _repository.AddAsync(request, cancellationToken);
-            return entity.Id;
-        }
-    }
-}
