+++ src/Pet.Jira.Infrastructure/ServiceCollectionExtensions.cs
@@ -1,8 +1,8 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Pet.Jira.Application.Articles;
 using Pet.Jira.Application.Authentication;
-using Pet.Jira.Application.Blog;
 using Pet.Jira.Application.Issues;
 using Pet.Jira.Application.Storage;
 using Pet.Jira.Application.Users;
@@ -10,8 +10,8 @@ using Pet.Jira.Application.Worklogs;
 using Pet.Jira.Application.Worklogs.Dto;
 using Pet.Jira.Domain.Models.Issues;
 using Pet.Jira.Domain.Models.Users;
+using Pet.Jira.Infrastructure.Articles;
 using Pet.Jira.Infrastructure.Authentication;
-using Pet.Jira.Infrastructure.Blog;
 using Pet.Jira.Infrastructure.Data.Contexts;
 using Pet.Jira.Infrastructure.Jira;
 using Pet.Jira.Infrastructure.Jira.Health;
@@ -62,7 +62,7 @@ namespace Pet.Jira.Infrastructure
             builder
                 .AddJiraHealthCheck()
                 .AddProcessAllocatedMemoryHealthCheck(
-                    maximumMegabytesAllocated: 200,
+                    maximumMegabytesAllocated: 300,
                     tags: new[] { "system" });
             return builder;
         }
