+++ src/Pet.Jira.Web/Components/Project/Articles/Article.razor.cs
@@ -1,5 +1,5 @@
 ﻿using Microsoft.AspNetCore.Components;
-using Pet.Jira.Application.Blog.Dto;
+using Pet.Jira.Application.Articles.Dto;
 
 namespace Pet.Jira.Web.Components.Project.Articles
 {
