+++ src/Pet.Jira.Web/Components/Worklogs/ActualWorklogItem.razor
@@ -1,15 +1,24 @@
 ﻿<MudElement Class="gap-1 d-flex flex-wrap border-none align-center">
     <MudElement Class="flex-auto gap-1 align-center" Style="width: 300px">
-        <WorklogTimeRange StartDate="@Entity.StartDate" CompleteDate="@Entity.CompleteDate" Color="Color.Tertiary" />
         <WorklogMenu Entity="@Entity" Color="Color.Tertiary" />
-        <MudLink Href="@Entity.Issue.Link" Typo="Typo.body2">@Entity.Issue.Key</MudLink> @Entity.Issue.Summary
+        <WorklogTimeRange StartDate="@Entity.StartDate" CompleteDate="@Entity.CompleteDate" Color="Color.Tertiary" />        
+        <MudElement Class="pl-2">
+            <MudButton Href="@Entity.Issue.Link"
+                       Size="Size.Small"
+                       Target="_blank"
+                       Variant="Variant.Text"
+                       Color="Color.Primary">
+                @Entity.Issue.Key
+            </MudButton>
+            @Entity.Issue.Summary
+        </MudElement>
     </MudElement>
     <MudElement Class="flex-none d-flex justify-end pet-time-spent-container align-center">
         <MudElement Class="flex-none d-flex gap-4" Style="width: 128px">
             <MudTextField @bind-Value="@Entity.RemainingTimeSpent" Disabled="true" Class="pet-input-control" Variant="Variant.Text" IconSize="Size.Small"></MudTextField>
             <MudElement Style="width: 40px">
                 <MudFab Size="Size.Small"
-                        Icon="@Icons.Material.Filled.Check"
+                        StartIcon="@Icons.Material.Filled.Check"
                         Color="Color.Tertiary" />
             </MudElement>
         </MudElement>
