+++ src/Pet.Jira.Web/Components/Worklogs/WorklogDayChip.razor
@@ -1,5 +1,9 @@
-﻿<MudChip Color="@Color" Size="Size.Small">
-    <b class="pet-not-display-on-480">@Caption: </b> @Value
+﻿<MudChip Color="@Color" Size=@Size Variant=@Variant Label="true">
+    @if (!string.IsNullOrEmpty(@Caption))
+    {
+        <b class="pet-not-display-on-480">@Caption: </b>
+    }
+    @Value
 </MudChip>
 
 @code {
