+++ src/Pet.Jira.Web/Components/Worklogs/WorklogItem.razor
@@ -3,9 +3,18 @@
 
 <MudElement Class="gap-1 d-flex flex-wrap border-none align-center">
     <MudElement Class="flex-auto gap-1 align-center" Style="width: 300px">
-        <WorklogTimeRange StartDate="@Entity.RawStartDate" CompleteDate="@Entity.RawCompleteDate" Color="@Color" />
         <WorklogMenu Entity="@Entity" Color="@Color" />
-        <MudLink Href="@Entity.Issue.Link" Typo="Typo.body2">@Entity.Issue.Key</MudLink> @Entity.Issue.Summary
+        <WorklogTimeRange StartDate="@Entity.RawStartDate" CompleteDate="@Entity.RawCompleteDate" Color="@Color" />
+        <MudElement Class="pl-2">
+            <MudButton Href="@Entity.Issue.Link"
+                       Size="Size.Small"
+                       Target="_blank"
+                       Variant="Variant.Text"
+                       Color="Color.Primary">
+                @Entity.Issue.Key
+            </MudButton>
+            @Entity.Issue.Summary
+        </MudElement>
     </MudElement>
     <MudElement Class="flex-none d-flex justify-end pet-time-spent-container align-center">
         <MudElement Class="flex-none d-flex gap-4" Style="width: 128px">
