+++ src/Pet.Jira.Web/Components/Worklogs/WorklogTimeRange.razor
@@ -1,4 +1,4 @@
-﻿<MudChip Color="@Color" Size="Size.Small">
+﻿<MudChip Color="@Color" Size="Size.Small" Variant="Variant.Filled" Label="true">
     @StartDate.ToString(_defaultTimeFormat) <MudIcon Icon="@Icons.Material.Filled.ArrowRightAlt" Size="Size.Small" /> @CompleteDate.ToString(_defaultTimeFormat)
 </MudChip>
