+++ src/Pet.Jira.Web/Shared/MainLayout.razor
@@ -20,9 +20,14 @@
             <MudIconButton Icon="@Icons.Material.Filled.Menu" Color="Color.Inherit" OnClick="@ToggleDrawer" />
             <MudText Typo="Typo.h6" Class="ml-3 pet-not-display-on-480">Jira workflow</MudText>
             <MudIconButton Icon="@Icons.Custom.Brands.GitHub" Color="Color.Inherit" Href="https://github.com/tolmachev-pravo/Pet.Jira" Target="_blank" />
-            <MudSwitch CheckedChanged="@ToggleThemeAsync" Checked="@_model.Theme.IsDarkMode" Color="Color.Primary" Class="ma-4" T="bool">
-                <MudElement Class="pet-not-display-on-480">Light/Dark</MudElement>
-            </MudSwitch>
+            <MudToggleIconButton Toggled="@_model.Theme.IsDarkMode"
+                                 Icon="@Icons.Material.Filled.LightMode"
+                                 Title="Light mode"
+                                 Color="@Color.Inherit"
+                                 ToggledIcon="@Icons.Material.Filled.DarkMode"
+                                 ToggledTitle="Dark mode"
+                                 ToggledColor="@Color.Inherit"
+                                 ToggledChanged="@ToggleThemeAsync"/>
             <MudSpacer></MudSpacer>
             <AuthorizeView>
                 <Authorized>
